comit

+++ PS.data/Infrastructure/IRepositoryBase.cs
@@ -11,7 +11,7 @@ namespace PS.Data.Infrastructure
         void Update(T entity);
         void Delete(T entity);
         void Delete(Expression<Func<T, bool>> where);
-        T GetById(long Id);
+        T GetById(int Id);
         T GetById(string Id);
         T Get(Expression<Func<T, bool>> where);
         IEnumerable<T> GetAll(); // GetMany()
