+++ PS.data/Infrastructure/RepositoryBase.cs
@@ -36,7 +36,7 @@ namespace PS.Data.Infrastructure
             foreach (T obj in objects)
                 dbset.Remove(obj);
         }
-        public virtual T GetById(long id)
+        public virtual T GetById(int id)
         {
             return dbset.Find(id);
         }
